+++ src/components/Wrapper.tsx
@@ -1,7 +1,7 @@
 'use client';
 import styles from './Wrapper.module.scss';
 import { ReactNode, Suspense, useEffect } from 'react';
-import { ReadonlyURLSearchParams, useRouter, usePathname, useSearchParams } from 'next/navigation';
+import { useRouter, usePathname, useSearchParams } from 'next/navigation';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks';
 
 import Header from './Header';
@@ -10,7 +10,6 @@ import { fetchSettings } from '@/modules/settings';
 import { autoLogin } from '@/modules/login';
 import Footer from './Footer';
 
-import { toast } from 'react-toastify';
 import { type Action } from '@reduxjs/toolkit';
 import { Permission } from '@/types';
 import Loading from '@/app/loader';
@@ -19,10 +18,6 @@ import { fetchTournaments } from '@/modules/tournament';
 import { fetchPartners } from '@/modules/partners';
 import { fetchAllCarts } from '@/modules/carts';
 
-interface SearchParams extends ReadonlyURLSearchParams {
-  action?: string;
-  state?: string;
-}
 /**
  * The navigation events component that is used to track navigation events.
  * Used mainly by Matomo
@@ -71,7 +66,6 @@ export default function Wrapper({
   children: ReactNode;
 }) {
   // Import necessary hooks and modules
-  const query: SearchParams = useSearchParams();
   const pathname = usePathname();
   const dispatch = useAppDispatch();
 
@@ -131,42 +125,6 @@ export default function Wrapper({
       }
     }
   }, [isLoggedIn, isLoginAllowed, isShopAllowed, isAdmin, isSpectator, hasTeam, pathname, isLoading]);
-
-  // TODO: implement a special route for the oauth callback
-  useEffect(() => {
-    if (isLoading) {
-      return;
-    }
-    // 1 action possible :
-    //  - oauth
-    if (query.action === 'oauth') {
-      switch (query.state) {
-        case '0':
-          toast.success('Le lien avec le compte Discord a bien été créé !');
-          toast.success('Tes rôles Discord te seront bientôt attribués !');
-          break;
-        case '1':
-          toast.success('Le lien avec le compte Discord a bien été mis à jour !');
-          toast.success('Tes rôles Discord te seront bientôt attribués !');
-          break;
-        case '2':
-          toast.success("Le lien avec le compte Discord n'a pas été modifié");
-          break;
-        case '3':
-          toast.error("Ce compte Discord est déjà lié au compte d'un autre utilisateur");
-          break;
-        case '4':
-          toast.error("Tu as refusé à nos services l'accès à ton compte Discord");
-          break;
-        case '5':
-          toast.error('Une erreur de requête est survenue');
-          break;
-        case '6':
-          toast.error('Une erreur inconnue est survenue');
-          break;
-      }
-    }
-  }, [isLoading]);
 
   const tournaments = useAppSelector((state) => state.tournament.tournaments);
   const partners = useAppSelector((state) => state.partners.partners);
